fix: list, show work
