+++ ros2diagnostics/ros2diagnostics/command/diagnostics.py
@@ -31,11 +31,8 @@ from ros2cli.command import add_subparsers_on_demand
 from ros2cli.command import CommandExtension
 
 
-class DiagCommand(CommandExtension):
-
-    def __init__(self):
-        super(DiagCommand, self).__init__()
-        self._subparser = None
+class DiagnosticsCommand(CommandExtension):
+    """Various diagnostics related sub-commands."""
 
     def add_arguments(self, parser, cli_name):
         self._subparser = parser
@@ -46,7 +43,7 @@ class DiagCommand(CommandExtension):
             'ros2diagnostics.verb',
             required=False)
 
-    def main(self, *, _, args):
+    def main(self, *, parser, args):
         if not hasattr(args, '_verb'):
             self._subparser.print_help()
             return 0
